requirements.txt
# CGS — Dependencies
# ─────────────────────────────────────────────
# Network & packet analysis
scapy>=2.6               # Python network reference — MIT — github.com/secdev/scapy
python-nmap>=0.7          # Nmap interface — pypi.org/project/python-nmap
# Web & real-time
flask>=3.0
flask-socketio>=5.3
gevent>=24.2
gevent-websocket>=0.10
# Infrastructure
pyyaml>=6.0
psutil>=5.9
schedule>=1.2
bcrypt>=4.1
peewee>=3.17
requests>=2.31
# CLI
click>=8.1
rich>=13.7
python-dateutil>=2.8
# PDF & DOCX (compliance and complaint reports)
reportlab>=4.0
python-docx>=1.1
cryptography>=41.0
markupsafe>=2.1
